feat: clean up unused imports and interfaces (#1838)
* feat: clean up unused imports and interfaces

* doc: javadoc fixes

* revert

* checkstyle

* style: line break changes

* cleanup

* Update ProcessSystemRunTransientTest.java

* Update TransferModelEnumsTest.java

+++ src/main/java/neqsim/thermodynamicoperations/flashops/ImprovedVUflashQfunc.java
@@ -142,7 +142,7 @@ public class ImprovedVUflashQfunc extends Flash {
     double initialPres = oldPres;
     double initialTemp = oldTemp;
 
-    double iterations = 0;
+    int iterations = 0;
     double maxIterations = 100; // Reduced for stability
     double tolerance = 1e-8; // Relaxed tolerance
