+++ src/main/java/neqsim/thermodynamicoperations/flashops/PHflash.java
@@ -74,7 +74,7 @@ public class PHflash extends Flash {
   public double solveQ() {
     double oldTemp = 1.0 / system.getTemperature();
     double nyTemp = 1.0 / system.getTemperature();
-    double iterations = 1;
+    int iterations = 1;
     double error = 1.0;
     double errorOld = 1.0e10;
     double factor = 0.8;
@@ -142,7 +142,7 @@ public class PHflash extends Flash {
   public double solveQ2() {
     double oldTemp = 1.0 / system.getTemperature();
     double nyTemp = 1.0 / system.getTemperature();
-    double iterations = 1;
+    int iterations = 1;
     double error = 1.0;
     double errorOld = 1.0e10;
     double factor = 0.8;
